backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pdfplumber==0.10.3
Werkzeug==2.3.7
pytesseract==0.3.10
Pillow==10.0.0
google-generativeai==0.3.0
python-dotenv==1.0.0
